+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-### 0.1.1-pre
+### 0.1.2-pre
+*
+
+### 0.1.1
 * Add group check attribute
 * Add headers and body as http(s) debug info
 * Fix exit code when all checks are successful